pyproject.toml
[project]
name = "uart-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "pyserial>=3.5",
    "unicorn>=2.1.4",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
